pyproject.toml
[project]
name = "llmoid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "fastapi>=0.128.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.40.0",
]
